refactor(entities)!: move E53_PlaceForm to `apis_core.entities`
BREAKING-CHANGE: We move `apis_core.apis_entities.forms.E53_PlaceForm`
to `apis_core.entities`, together with all the dependends. This affects:
* `fields.PlaceLookupField`
* `widgets.PlaceLookup`
* `js/placelookup.js`
* `templates/apis_entities/widgets/placelookup.html`
* `featureCodes_en.txt`
* the `get_feature_codes()` method in `apis_core.apis_entities.utils`
