--- apis_core/apis_entities/locale/de/LC_MESSAGES/django.po
@@ -1,31 +1,0 @@
-# SOME DESCRIPTIVE TITLE.
-# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER
-# This file is distributed under the same license as the PACKAGE package.
-# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.
-#
-#, fuzzy
-msgid ""
-msgstr ""
-"Project-Id-Version: PACKAGE VERSION\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2026-06-15 00:56-0500\n"
-"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
-"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
-"Language-Team: LANGUAGE <LL@li.org>\n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
-
-#: apis_core/apis_entities/forms.py
-msgid "Search places"
-msgstr "Orte suchen"
-
-#: apis_core/apis_entities/forms.py
-msgid "Click on a result to fill the form"
-msgstr "Ergebnis auswählen um die Daten zu übernehmen"
-
-#: apis_core/apis_entities/templates/apis_entities/widgets/placelookup.html
-msgid "Search"
-msgstr "Suchen"
