+++ apis_core/entities/featureCodes_en.txt
