+++ apis_core/entities/locale/de/LC_MESSAGES/django.po
@@ -86,6 +86,14 @@ msgstr "Gruppe"
 msgid "groups"
 msgstr "Gruppen"
 
+#: apis_core/entities/forms.py
+msgid "Search places"
+msgstr "Orte suchen"
+
+#: apis_core/entities/forms.py
+msgid "Click on a result to fill the form"
+msgstr "Ergebnis auswählen um die Daten zu übernehmen"
+
 #: apis_core/entities/templates/base.html
 msgid "Entities"
 msgstr "Entitäten"
@@ -133,3 +141,7 @@ msgstr "... kann nicht importiert werden"
 #: apis_core/entities/templates/entities/partials/linked_open_data.html
 msgid "delete uri"
 msgstr "URI löschen"
+
+#: apis_core/entities/templates/entities/widgets/placelookup.html
+msgid "Search"
+msgstr "Suchen"
