+++ apis_core/entities/templates/entities/widgets/placelookup.html
