+++ apis_core/entities/widgets.py
@@ -2,7 +2,7 @@ from django.forms.widgets import Input
 
 
 class PlaceLookup(Input):
-    template_name = "apis_entities/widgets/placelookup.html"
+    template_name = "entities/widgets/placelookup.html"
 
     class Media:
         js = ["js/placelookup.js"]
